+++ src/components/AddressContainer/index.tsx
@@ -26,6 +26,7 @@ import { nyCounties, nyCities, createQueryParams } from "../../utils/utils";
 import useFormDataContext from "../../context/FormDataContext";
 import { commonAPIErrors } from "../../data/apiErrorMessageTranslations";
 import { NRError } from "../../logger/newrelic";
+import { Paragraph } from "../Paragraph";
 
 const AddressContainer = ({ csrfToken }) => {
   const [isLoading, setIsLoading] = useState(false);
@@ -145,8 +146,8 @@ const AddressContainer = ({ csrfToken }) => {
 
   return (
     <>
-      <Heading level="three">{t("location.address.title")}</Heading>
-      <p>{t("location.address.description")}</p>
+      <Heading level="h3">{t("location.address.title")}</Heading>
+      <Paragraph>{t("location.address.description")}</Paragraph>
 
       <Loader isLoading={isLoading} />