+++ src/components/UsernameValidationFormFields/index.tsx
@@ -1,4 +1,5 @@
 import {
+  Box,
   Button,
   ButtonGroup,
   FormField as DSFormField,
@@ -13,7 +14,6 @@ import { isAlphanumeric } from "validator";
 
 import FormField from "../FormField";
 import useFormDataContext from "../../context/FormDataContext";
-import styles from "./UsernameValidationFormFields.module.css";
 
 import {
   apiErrorTranslations,
@@ -117,10 +117,6 @@ const UsernameValidationForm = ({
   };
   const inputValidation = (value = "") =>
     value.length >= 5 && value.length <= 25 && isAlphanumeric(value);
-  const availableClassname = usernameIsAvailable.available
-    ? styles.usernameAvailable
-    : styles.usernameUnavailable;
-
   /**
    * renderButton
    * Render the button to validate a username and enable it only if
@@ -171,9 +167,16 @@ const UsernameValidationForm = ({
       {usernameIsAvailable?.message ? (
         <FormRow id={`${id}-username-3`}>
           <DSFormField>
-            <div className={availableClassname} aria-live="assertive">
+            <Box
+              color={
+                usernameIsAvailable.available
+                  ? "var(--nypl-colors-ui-success-primary)"
+                  : "var(--nypl-colors-ui-error-primary)"
+              }
+              aria-live="assertive"
+            >
               {usernameIsAvailable.message}
-            </div>
+            </Box>
           </DSFormField>
         </FormRow>
       ) : null}